In the World Clock section, display Local time like in the alarm page. Then add the time for the following popular cities "New York, Chicago, Denver, Los Angeles, Phoenix, Anchorage, Honolulu, Toronto, London, Sydney, Manila, Singapore, Tokyo, Beijing, Berlin, Mexico City, Buenos Aires, Dubai".
Make each city name clickable so that when a user clicks on it, it will open a new page showing the time in that city.

For seo purposes, add a section under the city time section for "Current time in City". In this section display the time zone of the city and a brief information about the city in one paragraph.

Add another section under it for "What time is it in City right now?" In this section, use the information about the time of the city to two short paragraphs about the time in the city at that time. You can add what season it is from the date information.